+++ src/modules/files/files.module.ts
@@ -3,11 +3,12 @@ import { FilesController } from './files.controller.js';
 import { FilesService } from './files.service.js';
 import { StorageModule } from '../storage/storage.module.js';
 import { OptimizationModule } from '../optimization/optimization.module.js';
+import { ExifService } from './exif.service.js';
 
 @Module({
   imports: [StorageModule, OptimizationModule],
   controllers: [FilesController],
-  providers: [FilesService],
+  providers: [FilesService, ExifService],
   exports: [FilesService],
 })
 export class FilesModule {}